+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## dev
+
+### Changed
+
+ * Always read task names as lowercase when parsing metadata.csv files ([#95])
+ * Parse fastq.gz files per sample from disk rather than predicting from sample
+   name alone ([#94])
+
+### Fixed
+
+ * Corrected parsing for Illumina's Checkpoint.txt files to work with
+   incomplete alignment outputs ([#97])
+
+[#97]: https://github.com/ShawHahnLab/umbra/pull/97
+[#95]: https://github.com/ShawHahnLab/umbra/pull/95
+[#94]: https://github.com/ShawHahnLab/umbra/pull/94
+
 ## 0.0.3 - 2020-08-19
 
 ### Added